DomainBed/domainbed/requirements.txt
numpy==1.24.1
wilds==2.0.0
imageio==2.9.0
gdown==4.6.0
torchvision==0.14.1+cu116
torch==1.13.1+cu116
tqdm==4.64.1
backpack==0.1
parameterized==0.8.1
Pillow==9.4.0
pip install cuml-cu12 --extra-index-url=https://pypi.nvidia.com